Not uploading a golang tool's dependencies can cause problems sometimes.
Unlike node or java apps, Go puts the burden on end users to “go get”
dependencies themselves, or as of go 1.5 uses vendoring.  I initially
thought the vendoring would allow me to avoid uploading all the
dependencies to github, but tools that depend on this tool may not have
this tool’s dependencies downloaded, so vending is preferred.