Apply golang standads for file organization in a library.
Golang standards dictate that tools should contain a main package and main function, but libraries should not have a main package or function. This commit updates terratest to be treated like a proper library. It places unit tests in the same package and integration tests into terratest_test.go at the top-level. It also replaces the main package with the "terratest" package.