Create a cleaner interface (set of methods) for library consumers to use.
Previously, terratest had simply an “Apply” method which not only
applied but also destroyed. In some cases, other libraries may wish to
use terratest to provision test resources in AWS (and therefore not
immediately destroy the resources).

For these reasons, a new set of methods “Apply”, “ApplyAndDestroy” and
“Destroy” have been introduced.  In addition, code has been organized
to be cleaner, and tests have been cleaned up.