Exclude regions from specific terraform apply runs. Also some clean-up to code base.
Previously, we could globally tell terratest never to run tests in a
particular region. But this became a limitation when we wished to test
the VPC, and the template contained a NAT Gateway resource which is not
yet available in all AWS regions.

Now each test can declare which regions it cannot be run in. This is
useful both to avoid false negatives in our testing and also to declare
to customers for which regions a given template is supported.

For the code cleanup, some duplication was found with respect to
RandomResourceCollections and that has now been removed.  In addition,
ApplyAndDestroy unnecessarily copied code, whereas now it relies on
Apply only (credit to Jim for this!).