Add URL checker
Add terratest helpers for checking URLs. The helpers will hit a URL
repeatedly until it returns a 200 OK with the expected body or they
reach the maximum number of retries and return an error. This is useful
for testing that a server or set of servers behind an ELB are really up
and running. Also included are unit tests for these helpers that verify
the helpers work correctly by running a dummy HTTP server.