Fetch the default VPC instead of a random one
We should not be using VPCs at random. Our tests create and destroy
VPCs all the time, so it's possible that one test will accidentally try
to use a VPC that another test is trying to destroy. This changes
terratest to return the default VPC instead of a random one.