Preserve folder name when copying
The `CopyTerraformFolderToTemp` method originally copied everything
into a folder with a random name. I’ve now updated the method to still
use a randomly named temp folder (so it doesn’t clash with other
folders), but to copy everything to a folder within the temp folder
that uses the same name as the source. This makes it easier to
understand what’s going on when debugging. Moreover, some of our code
relies on that original folder name, such as the way we do Packer
builds for open source projects (e.g., consul, nomad, etc.).