Add support for saving and loading multiple AMIs.
While working with the Confluent and Kafka tests, I had to build in some cases up to 3 separate AMIs to run a single test. The previous test_structucture code onlyh allowed for saving and loading a single AMI, so this adds a backwards-compatible way of saving and loading multiple AMIs at once.