Fix output commands to only read from stdout
The `Output` commands in Terratest were running `terraform output` and then parsing the results from both `stdout` and `stderr`. This works with Terraform, as its `output` commands solely right to `stdout`, though it's a risky way to do it and may well break in the future; more to the point, it's broken with `terragrunt output` commands. Added failing test case and then fixed the code to make the test case pass.