Add guide on what types of contributions we accept in Terratest
We've started to get a number of small contributions that feel like esoteric use cases or very thin wrappers for cloud APIs. To prevent Terratest from becoming a giant, bloated library that mirrors the entire AWS SDK, GCP SDK, etc, I've updated the contribution guidelines with what types of helper functions we're interested in seeing.

I don't want to scare off potential contributions, of course, so I'm especially grateful for feedback on how to balance this guide in such a way that the community feels comfortable in continuing to add awesome features, but without Terratest becoming unusable.