Move test to normal build
Because the logparser could not handle really long lines, a test had to
be moved behind build tags so that it does not get executed with the
"normal" test suite, but instead with one where the logger is disabled.
As the logger is using the new ReadLine now, this test can be moved into
the usual test file.

Signed-off-by: Ramon Rüttimann <me@ramonr.ch>