Improve logging facilities
This commit introduces a better logging facility. It adds some types to
allow the caller to customise logging, while not introducing
API-breaking changes.

It enables the user to disable logging for terraform commands and shell
commands. The user may also supply his own Logf function (or even
multiple Logf functions) to these packages to further customise the
output.

The default Logf function is kept and has changed to use testing.T.Logf
if the Go version is at least 1.14. This is because with Go 1.14,
streaming output for Logf has been introduced. The inability to print
the output during testing was why the logger package was created in the
first place. This is fixed now, so if the logger package can determine
the Go version and that version supports streaming Logf, it will be used
instead.

Furthermore, some default loggers have been created:

- Discard:  discards all logging. Useful for functions where the output
            is too large or or unnecessary.
- TestingT: uses testing.T, if supplied (falls back to the "old" logger
            if not). This means that users with pre-1.14 versions can
            use testing.T.Logf anyway, if they want to.
- Logger:   This is the "old" logger that has been used so far. Can be
            used if, for example, the go version is >=1.14, but the
            "default" testing.T.Logf should not be used for whatvever
            reason.

This is all implemented while maintaining the old API. For users of that
API, nothing changes.

Signed-off-by: Ramon Rüttimann <me@ramonr.ch>